Fix for CPU usage ignored by idler
PROBLEM(S)
==========

1) First problem was that `metrics_lookup` dictionary keys were build
off pods' names but lookup was done by deployment name
2) Second problem was that `avg_cpu_percent()` was summing the number
of used millicpus but that's not actually the percentage of CPU usage

SOLUTION(S)
===========

Solution to first problem was to use the `(app_name, namespace)` as key
(see caviat below).

The solution to the second problem is to use the following logic to determine
the percentage of a deployment CPU used.

We still sum the CPU usage for all containers in the pod.
We also sum the CPU limits for all containers in the pod (from the deployement
template).

Then we divide the two, and that's the percentage of used CPU.

CAVIAT
======

This is mainly targeting hidling of RStudio instances.
We're assuming here that there is:
1) One instance of RStudio per user/namespace
2) One replica/pod for each of instance of RStudio

When/if this will change the logic to determine whether a deployment
is in use will need to get more complicated as it will need to group
by deployment rather than by app name.

TICKET
======

https://trello.com/c/jgHhwqcF/736-fix-bug-with-hibernations-idling-so-that-users-can-run-jobs-overnight