Lookup Ingress by app label
This looks up the ingress by app label, to support jupyter-lab deployments which
call their ingress object something that doesn't match the deployment name.
Looking up by the app label is safe because this is a required label for tools
and is used to show the app name in the cpanel so we can be sure that all tools
will have an app label.