Moooaar logging. We want to know more, not less
There are some cases where idled applications are in an inconsistent
state. It's not clear why/how that happened but I'm adding more logging
so that if something strange happens we can see the logs and have a
better understanding of which steps are performed correctly and which are not.