Better exception handing
Handle all exceptions and log an error instead of causing an exception
to stop the loop and prevent idling of subsequent deployments.

The list of Deployments which were not idled is then shown at the end.
If if was not possible to idle any of the Deployments the idler also
exits with a non-zero value so that the job is marked as failed as it
should in this case.