Patch Service using strategic merge instead of JSONPatch
JSONPatch is been a source of troubles. Specifically because of the
fact that the specification says that the patch should not work
when trying to remove a key which is not present.
Even "worse" the whole patch will be rejected so none of the changes
would be applied.

Basically a JSONPatchs are not applied in an idempotent way as we thought.

Strategic merge seems more predictable/declarative and it doesn't blow up
when the key is not present already.

**NOTE**: To remove a key from a dictionary/map it's possible to set that
key to `null`.