Simplified Deployment patch
We used to have several methods which modified the `Deployment`
object in memory and then write these changes by calling patch with the
whole object.

This was a bit misleading as it appear by reading the logs/code that
something was changing in k8s while in reality nothing was until the
final "write_changes()". It also didn't help that the service was
redirected in the middle of these deployment operations.

I've moved the changes in a single (smaller) patch into a new class
called `App` which is both redirecting the `Service` and scale down/update
the `Deployment`.

The idler code is still not ideal as it's a mixture of procedural and
object-oriented code but I don't want to change too many things at once
as it's essentially working.